feat(core/util): extract shared parseIntOr/parseFloatOr helpers (BEC-188 pre-req)
Creates packages/core/src/util/env.ts with NaN-safe numeric parsing helpers
(parseIntOr, parsePosIntOr, parseFloatOr, parseOptPosInt) and exports them
from @urateam/core. Replaces three local copies in review-providers-runner.ts,
review-provider.ts, and github-from-env.ts. Fixes unguarded parseInt for
GITHUB_INSTALLATION_ID (could silently produce NaN).
