+++ packages/core/src/index.ts
@@ -1,4 +1,5 @@
 // SPDX-License-Identifier: BUSL-1.1
+export { parseIntOr, parsePosIntOr, parseFloatOr, parseOptPosInt } from "./util/env.js";
 export * from "./types.js";
 export {
   checkLicense,
